komunikat

+++ src/com/sda/oajiety/InerfaceTest.java
@@ -1,6 +1,6 @@
 package com.sda.oajiety;
 
 public interface InerfaceTest {
-    int pole ();
-    int obwod ();
+    double pole ();
+    double obwod ();
 }
